Fix AN/ASN-128 stuck knob
This is not a fully functional fix due to a few edge cases, but the knob will no longer end up completely stuck. This breaks down slightly when the display knob is set to position 3 (PP), and the mode switch is moved to position 0 (off) then to position 3, then not pressing enter to make the system operational. In this specific configuration using the keybind to increase the display knob will fail because displayIndex = 100 (increase keybind requres displayIndex < 60). Using the keybind to decrease the knob will snap the knob to position 6. Fully resolving this will probably require some amount of rewrite, or a clever solution that I am unable to think of. In the meantime this is workable.