Improve HSI km Readout
Bypassed the moveGauge animation speeds and implemented the same animation logic in the actual range commanded to be displayed on the gauge. This provides proper animation where all digits must be scrolled though for large changes (instead of having digits independently move), as well as fixes the peculiar "unwinding" of the digits as they transit between 9 and 0 (and vice versa).