Fix momentary display glitch
A brief but noticeable display glitch would occur when moving the switch from off to test. Whatever was on the display before moving to off would briefly appear before the test procedure ran. This impacted cold starts (it would display NULL) as well as any time the switch was moved off in the middle of the self test.