Remove dofile for Curve.lua
It causes an error and doesn't appear to be needed. I think perhaps the person who posted the solution was erroneously under the impression that Curve.lua was part of the fix when it's in fact just another part of the A-4 mod, and is unrelated to the UH-60L or the sound issue.